feat: add basic router logic
