+++ package.json
@@ -20,7 +20,8 @@
   "dependencies": {
     "@tanstack/react-query": "^5.81.5",
     "react": "^19.1.0",
-    "react-dom": "^19.1.0"
+    "react-dom": "^19.1.0",
+    "react-router": "^7.7.1"
   },
   "devDependencies": {
     "@eslint/js": "^9.30.1",
@@ -36,7 +37,7 @@
     "eslint-plugin-prettier": "^5.5.1",
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-react-refresh": "^0.4.20",
-    "eslint-plugin-unicorn": "^59.0.1",
+    "eslint-plugin-unicorn": "^60.0.0",
     "gh-pages": "^6.3.0",
     "globals": "^16.3.0",
     "husky": "^9.1.7",
