+++ src/pages/about/about.tsx
@@ -1,0 +1,5 @@
+import type { JSX } from 'react';
+
+export default function About(): JSX.Element {
+  return <h1>About Page</h1>;
+}
