+++ src/pages/no-page/no-page.tsx
@@ -1,0 +1,5 @@
+import type { JSX } from 'react';
+
+export default function NoPage(): JSX.Element {
+  return <h1>404 Page</h1>;
+}
